backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.3.0
librosa==0.10.2.post1
soundfile==0.13.1
torch==2.8.0
torchaudio==2.8.0
transformers==4.55.4
speechbrain==1.0.3
tensorflow==2.19.0
tf-keras==2.19.0
fer==22.5.1
moviepy==1.0.3
